requirements.txt
fastapi
uvicorn
pytest
ruff
httpx
# Database
sqlalchemy>=2.0
psycopg[binary]>=3.1

# Config
pydantic-settings>=2.0
yfinance 
pandas
joblib
scikit-learn
optuna
shap
sentence-transformers
faiss-cpu
